Refactor fluent integers
