+++ parasol_runtime/Cargo.toml
@@ -20,6 +20,7 @@ mux_circuits = { workspace = true }
 num = { workspace = true }
 parasol_concurrency = { workspace = true }
 petgraph = { workspace = true }
+rand = { workspace = true }
 rayon = { workspace = true }
 reqwest = { version = "0.12.15", features = ["blocking"] }
 serde = { workspace = true }
@@ -35,7 +36,6 @@ clap = { workspace = true }
 criterion = { workspace = true }
 indicatif = { workspace = true }
 ndarray = "0.16.1"
-rand = { workspace = true }
 scirs2-optimize = "0.1.0-alpha.3"
 sunscreen_math = { workspace = true }
