+++ parasol_runtime/src/fluent/int.rs
@@ -1,10 +1,10 @@
-use crate::{L1GlweCiphertext, circuits::mul::append_int_multiply};
+use crate::{circuits::mul::append_int_multiply, fluent::{DynamicGenericInt, PlaintextOps, GenericIntGraphNodes, PackedGenericIntGraphNode}, L1GlweCiphertext};
 
 use super::{
     FheCircuit, Muxable, PackedGenericInt,
     generic_int::{
-        DynamicGenericInt, GenericInt, GenericIntGraphNodes, PackedDynamicGenericInt,
-        PackedGenericIntGraphNode, Sign,
+        GenericInt, PackedDynamicGenericInt,
+        Sign,
     },
 };
 
@@ -16,7 +16,46 @@ use serde::{Deserialize, Serialize};
 #[derive(Clone, Serialize, Deserialize)]
 pub struct Signed;
 
+impl PlaintextOps for i128 {
+    fn assert_in_bounds(&self, bits: usize) {
+        assert!(*self < 0x1 << (bits - 1));
+        assert!(*self >= -(0x1 << (bits - 1)));
+    }
+
+    // fn extract_bit_as_bool(&self, bit: usize) -> bool {
+    //     self.extract_bit(bit) == 0x1
+    // }
+
+    // fn extract_bit(&self, bit: usize) -> Self {
+    //     (*self >> bit) & 0x1
+    // }
+
+    fn from_bits<I: Iterator<Item=bool>>(i: I) -> Self {
+        let mut bits = 0;
+
+        let mut val = i.enumerate().fold(0i128, |s, (i, x)| {
+            bits += 1;
+            s | ((x as i128) << i)
+        });
+
+        let sign = (val >> bits - 1) & 0x1;
+
+        // Sign extend our value to 128-bit.
+        for i in bits..128 {
+            val |= sign << i;
+        }
+
+        val
+    }
+
+    fn to_bits(&self, len: usize) -> impl Iterator<Item=bool> {
+        (0..len).enumerate().map(|(i, _)| ((*self >> i) & 0x1) == 0x1)
+    }
+}
+
 impl Sign for Signed {
+    type PlaintextType = i128;
+
     fn gen_compare_circuit(max_len: usize, gt: bool, eq: bool) -> mux_circuits::MuxCircuit {
         compare_or_maybe_equal_signed(max_len, gt, eq)
     }
@@ -95,6 +134,18 @@ pub type PackedInt128 = PackedInt<128, L1GlweCiphertext>;
 /// Encrypted packed 256 bit integer. This is a specialization of [`PackedInt`] for 256 bits and [`L1GlweCiphertext`].
 pub type PackedInt256 = PackedInt<256, L1GlweCiphertext>;
 
+/// Given input [`u128`] of `bits` length, sign extend and return the value as
+/// an [`i128`]
+pub(crate) fn sign_extend(mut val: u128, bits: u32) -> i128 {
+    let sign = (val >> (bits - 1)) & 0x1;
+
+    for i in bits..128 {
+        val += sign << i;
+    }
+
+    val as i128
+}
+
 #[cfg(test)]
 mod tests {
     use crate::{
@@ -114,9 +165,9 @@ mod tests {
         let sk = get_secret_keys_128();
         let pk = get_public_key_128();
 
-        let val = PackedInt::<16, L1GlweCiphertext>::encrypt(2u64.pow(16) - 42, &enc, &pk);
+        let val = PackedInt::<16, L1GlweCiphertext>::encrypt(-42, &enc, &pk);
 
-        assert_eq!(val.decrypt(&enc, &sk), 2u64.pow(16) - 42);
+        assert_eq!(val.decrypt(&enc, &sk), -42);
     }
 
     #[test]
@@ -126,9 +177,9 @@ mod tests {
         let sk = get_secret_keys_128();
         let pk = get_public_key_128();
 
-        let val = PackedDynamicInt::<L1GlweCiphertext>::encrypt(2u64.pow(16) - 42, &enc, &pk, 16);
+        let val = PackedDynamicInt::<L1GlweCiphertext>::encrypt(-42, &enc, &pk, 16);
 
-        assert_eq!(val.decrypt(&enc, &sk), 2u64.pow(16) - 42);
+        assert_eq!(val.decrypt(&enc, &sk), -42);
     }
 
     #[test]
@@ -139,7 +190,7 @@ mod tests {
         let pk = get_public_key_128();
         let (uproc, fc) = make_uproc_128();
 
-        let val = PackedInt::<16, L1GlweCiphertext>::encrypt(2u64.pow(16) - 42, &enc, &pk);
+        let val = PackedInt::<16, L1GlweCiphertext>::encrypt(-42, &enc, &pk);
 
         let ctx = FheCircuitCtx::new();
 
@@ -154,7 +205,7 @@ mod tests {
             .run_graph_blocking(&ctx.circuit.borrow(), &fc)
             .unwrap();
 
-        assert_eq!(as_unpacked.decrypt(&enc, &sk), 2u128.pow(16) - 42);
+        assert_eq!(as_unpacked.decrypt(&enc, &sk), -42);
     }
 
     #[test]
@@ -165,7 +216,7 @@ mod tests {
         let pk = get_public_key_128();
         let (uproc, fc) = make_uproc_128();
 
-        let val = PackedDynamicInt::<L1GlweCiphertext>::encrypt(2u64.pow(16) - 42, &enc, &pk, 16);
+        let val = PackedDynamicInt::<L1GlweCiphertext>::encrypt(-42, &enc, &pk, 16);
 
         let ctx = FheCircuitCtx::new();
 
@@ -180,7 +231,7 @@ mod tests {
             .run_graph_blocking(&ctx.circuit.borrow(), &fc)
             .unwrap();
 
-        assert_eq!(as_unpacked.decrypt(&enc, &sk), 2u128.pow(16) - 42);
+        assert_eq!(as_unpacked.decrypt(&enc, &sk), -42);
     }
 
     #[test]
@@ -189,7 +240,7 @@ mod tests {
         let sk = get_secret_keys_128();
         let (uproc, fc) = make_uproc_128();
 
-        let val = Int::<15, L1GlweCiphertext>::encrypt_secret(2u128.pow(15) - 42, &enc, &sk);
+        let val = Int::<15, L1GlweCiphertext>::encrypt_secret(-42, &enc, &sk);
 
         let ctx = FheCircuitCtx::new();
 
@@ -204,7 +255,7 @@ mod tests {
             .run_graph_blocking(&ctx.circuit.borrow(), &fc)
             .unwrap();
 
-        assert_eq!(actual.decrypt(&enc, &sk), 2u64.pow(15) - 42);
+        assert_eq!(actual.decrypt(&enc, &sk), -42);
     }
 
     #[test]
@@ -213,7 +264,7 @@ mod tests {
         let sk = get_secret_keys_128();
         let (uproc, fc) = make_uproc_128();
 
-        let val = DynamicInt::<L1GlweCiphertext>::encrypt_secret(2u128.pow(15) - 42, &enc, &sk, 15);
+        let val = DynamicInt::<L1GlweCiphertext>::encrypt_secret(-42, &enc, &sk, 15);
 
         let ctx = FheCircuitCtx::new();
 
@@ -228,7 +279,7 @@ mod tests {
             .run_graph_blocking(&ctx.circuit.borrow(), &fc)
             .unwrap();
 
-        assert_eq!(actual.decrypt(&enc, &sk), 2u64.pow(15) - 42);
+        assert_eq!(actual.decrypt(&enc, &sk), -42);
     }
 
     #[test]
@@ -237,7 +288,7 @@ mod tests {
             let enc = get_encryption_128();
             let sk = get_secret_keys_128();
 
-            let val = Int::<15, T>::encrypt_secret(2u128.pow(15) - 42, &enc, &sk);
+            let val = Int::<15, T>::encrypt_secret(-42, &enc, &sk);
 
             let ser = bincode::serialize(&val).unwrap();
             crate::safe_bincode::deserialize::<Int<15, T>>(&ser, &DEFAULT_128).unwrap();
@@ -255,7 +306,7 @@ mod tests {
             let enc = get_encryption_128();
             let sk = get_secret_keys_128();
 
-            let val = DynamicInt::<T>::encrypt_secret(2u128.pow(15) - 42, &enc, &sk, 15);
+            let val = DynamicInt::<T>::encrypt_secret(-42, &enc, &sk, 15);
 
             let ser = bincode::serialize(&val).unwrap();
             bincode::deserialize::<DynamicInt<T>>(&ser).unwrap();
@@ -273,7 +324,7 @@ mod tests {
         let sk = get_secret_keys_128();
         let pk = PublicKey::generate(&DEFAULT_128, &sk);
 
-        let val = PackedInt::<15, L1GlweCiphertext>::encrypt(2u64.pow(15) - 42, &enc, &pk);
+        let val = PackedInt::<15, L1GlweCiphertext>::encrypt(2i128.pow(15) - 42, &enc, &pk);
 
         let ser = bincode::serialize(&val).unwrap();
         crate::safe_bincode::deserialize::<PackedInt<15, L1GlweCiphertext>>(&ser, &DEFAULT_128)
@@ -286,7 +337,7 @@ mod tests {
         let sk = get_secret_keys_128();
         let pk = PublicKey::generate(&DEFAULT_128, &sk);
 
-        let val = PackedDynamicInt::<L1GlweCiphertext>::encrypt(2u64.pow(15) - 42, &enc, &pk, 15);
+        let val = PackedDynamicInt::<L1GlweCiphertext>::encrypt(2i128.pow(15) - 42, &enc, &pk, 15);
 
         let ser = bincode::serialize(&val).unwrap();
         crate::safe_bincode::deserialize::<PackedDynamicInt<L1GlweCiphertext>>(&ser, &DEFAULT_128)
@@ -298,9 +349,9 @@ mod tests {
         let enc = get_encryption_128();
         let sk = get_secret_keys_128();
 
-        let val = PackedInt::<15, L1GlweCiphertext>::trivial_encrypt(2u64.pow(15) - 42, &enc);
+        let val = PackedInt::<15, L1GlweCiphertext>::trivial_encrypt(-42, &enc);
 
-        assert_eq!(val.decrypt(&enc, &sk), 2u64.pow(15) - 42);
+        assert_eq!(val.decrypt(&enc, &sk), -42);
     }
 
     #[test]
@@ -309,8 +360,8 @@ mod tests {
         let sk = get_secret_keys_128();
 
         let val =
-            PackedDynamicInt::<L1GlweCiphertext>::trivial_encrypt(2u64.pow(15) - 42, &enc, 15);
+            PackedDynamicInt::<L1GlweCiphertext>::trivial_encrypt(-42, &enc, 15);
 
-        assert_eq!(val.decrypt(&enc, &sk), 2u64.pow(15) - 42);
+        assert_eq!(val.decrypt(&enc, &sk), -42);
     }
 }
