+++ parasol_runtime/src/fluent/uint.rs
@@ -1,10 +1,10 @@
-use crate::{L1GlweCiphertext, circuits::mul::append_uint_multiply};
+use crate::{circuits::mul::append_uint_multiply, fluent::{DynamicGenericInt, GenericIntGraphNodes, PackedGenericIntGraphNode, PlaintextOps}, L1GlweCiphertext};
 
 use super::{
     FheCircuit, Muxable, PackedGenericInt,
     generic_int::{
-        DynamicGenericInt, GenericInt, GenericIntGraphNodes, PackedDynamicGenericInt,
-        PackedGenericIntGraphNode, Sign,
+        GenericInt, PackedDynamicGenericInt,
+        Sign,
     },
 };
 
@@ -16,7 +16,33 @@ use serde::{Deserialize, Serialize};
 #[derive(Clone, Serialize, Deserialize)]
 pub struct Unsigned;
 
+impl PlaintextOps for u128 {
+    fn assert_in_bounds(&self, bits: usize) {
+        assert!(*self < 0x1 << bits);
+    }
+
+    // fn extract_bit_as_bool(&self, bit: usize) -> bool {
+    //     self.extract_bit(bit) == 0x1
+    // }
+
+    // fn extract_bit(&self, bit: usize) -> Self {
+    //     (*self >> bit) & 0x1
+    // }
+
+    fn from_bits<I: Iterator<Item=bool>>(iter: I) -> Self {
+        iter.enumerate().fold(0u128, |s, (i, x)| {
+            s + ((x as u128) << i)
+        })
+    }
+
+    fn to_bits(&self, len: usize) -> impl Iterator<Item=bool> {
+        (0..len).enumerate().map(|(i, _)| ((*self >> i) & 0x1) == 0x1)
+    }
+}
+
 impl Sign for Unsigned {
+    type PlaintextType = u128;
+
     fn gen_compare_circuit(max_len: usize, gt: bool, eq: bool) -> mux_circuits::MuxCircuit {
         compare_or_maybe_equal(max_len, gt, eq)
     }
